+++ packages/api-derive/src/chain/subscribeNewHeads.ts
@@ -5,10 +5,11 @@ import type { Observable } from 'rxjs';
 import type { HeaderExtended } from '../type/types';
 import type { DeriveApi } from '../types';
 
-import { combineLatest, map, of } from 'rxjs';
+import { combineLatest, map, of, switchMap } from 'rxjs';
 
 import { createHeaderExtended } from '../type';
 import { memo } from '../util';
+import { getAuthorDetails } from './util';
 
 /**
  * @name subscribeNewHeads
@@ -25,16 +26,20 @@ import { memo } from '../util';
  */
 export function subscribeNewHeads (instanceId: string, api: DeriveApi): () => Observable<HeaderExtended> {
   return memo(instanceId, (): Observable<HeaderExtended> =>
-    combineLatest([
-      api.rpc.chain.subscribeNewHeads(),
-      api.query.session
-        ? api.query.session.validators()
-        : of(undefined)
-    ]).pipe(
-      map(([header, validators]): HeaderExtended => {
+    api.rpc.chain.subscribeNewHeads().pipe(
+      switchMap((header) =>
+        combineLatest([
+          of(header),
+          api.queryAt(header.hash)
+        ])
+      ),
+      switchMap(([header, queryAt]) =>
+        getAuthorDetails(header, queryAt)
+      ),
+      map(([header, validators, author]): HeaderExtended => {
         header.createdAtHash = header.hash;
 
-        return createHeaderExtended(header.registry, header, validators);
+        return createHeaderExtended(header.registry, header, validators, author);
       })
     )
   );