+++ packages/api-derive/src/type/HeaderExtended.ts
@@ -7,19 +7,17 @@ import type { HeaderExtended } from './types';
 
 import { extractAuthor } from './util';
 
-export function createHeaderExtended (registry: Registry, header?: Header, validators?: AccountId[]): HeaderExtended {
+export function createHeaderExtended (registry: Registry, header?: Header, validators?: AccountId[] | null, author?: AccountId | null): HeaderExtended {
   // an instance of the base extrinsic for us to extend
   const HeaderBase = registry.createClass('Header');
 
   class Implementation extends HeaderBase implements HeaderExtended {
     readonly #author?: AccountId;
-    readonly #validators?: AccountId[];
 
-    constructor (registry: Registry, header?: Header, validators?: AccountId[]) {
+    constructor (registry: Registry, header?: Header, validators?: AccountId[] | null, author?: AccountId | null) {
       super(registry, header);
 
-      this.#author = extractAuthor(this.digest, validators);
-      this.#validators = validators;
+      this.#author = author || extractAuthor(this.digest, validators || []);
       this.createdAtHash = header?.createdAtHash;
     }
 
@@ -29,14 +27,7 @@ export function createHeaderExtended (registry: Registry, header?: Header, valid
     public get author (): AccountId | undefined {
       return this.#author;
     }
-
-    /**
-     * @description Convenience method, returns the validators for the block
-     */
-    public get validators (): AccountId[] | undefined {
-      return this.#validators;
-    }
   }
 
-  return new Implementation(registry, header, validators);
+  return new Implementation(registry, header, validators, author);
 }